backend/requirements.txt
Django==3.2.16
djangorestframework-simplejwt==5.3.1
gunicorn==20.0.4
python-dotenv==0.21.0
psycopg2-binary==2.9.5